phishguard/requirements.txt
# Core dependencies
python-dotenv>=0.19.0
requests>=2.26.0
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0
joblib>=1.1.0

# Web and API
fastapi>=0.70.0
uvicorn>=0.15.0
jinja2>=3.0.0
python-multipart>=0.0.5

# URL and HTML processing
beautifulsoup4>=4.10.0
urllib3>=1.26.7
tldextract>=3.1.2
python-whois>=0.8.0

# Machine Learning
tensorflow>=2.6.0
xgboost>=1.4.0

# Development
tox>=3.24.0
pytest>=6.2.5
black>=21.12b0
flake8>=4.0.1
